refactor: isolate translation constants
Keep Bun translation tooling off the React Native import graph while preserving shared runtime values and type contracts.
