+++ src/CONST/DEFAULT_NEW_EXPENSIFY_URL.ts
@@ -1,0 +1,3 @@
+const DEFAULT_NEW_EXPENSIFY_URL = 'https://new.expensify.com';
+
+export default DEFAULT_NEW_EXPENSIFY_URL;
