+++ src/CONST/NEW_EXPENSIFY_URL.ts
@@ -1,0 +1,5 @@
+import DEFAULT_NEW_EXPENSIFY_URL from './DEFAULT_NEW_EXPENSIFY_URL';
+
+const NEW_EXPENSIFY_URL = `${DEFAULT_NEW_EXPENSIFY_URL}/`;
+
+export default NEW_EXPENSIFY_URL;
