+++ src/CONST/REPORT_ACTION_TYPES.ts
@@ -1,0 +1,259 @@
+// cspell:disable
+const REPORT_ACTION_TYPES = {
+    ACTIONABLE_ADD_PAYMENT_CARD: 'ACTIONABLEADDPAYMENTCARD',
+    // Concierge message informing the user of a 3DS challenge - custom reportAction type because we want to translate it
+    ACTIONABLE_CARD_3DS_TRANSACTION_APPROVAL: 'ACTIONABLE_CARD_3DS_TRANSACTION_APPROVAL',
+    ACTIONABLE_CARD_FRAUD_ALERT: 'ACTIONABLECARDFRAUDALERT',
+    ACTIONABLE_JOIN_REQUEST: 'ACTIONABLEJOINREQUEST',
+    ACTIONABLE_MENTION_WHISPER: 'ACTIONABLEMENTIONWHISPER',
+    ACTIONABLE_MENTION_INVITE_TO_SUBMIT_EXPENSE_CONFIRM_WHISPER: 'ACTIONABLEMENTIONINVITETOSUBMITEXPENSECONFIRMWHISPER',
+    ACTIONABLE_REPORT_MENTION_WHISPER: 'ACTIONABLEREPORTMENTIONWHISPER',
+    ACTIONABLE_TRACK_EXPENSE_WHISPER: 'ACTIONABLETRACKEXPENSEWHISPER',
+    POLICY_EXPENSE_CHAT_WELCOME_WHISPER: 'POLICYEXPENSECHATWELCOMEWHISPER',
+    ADD_COMMENT: 'ADDCOMMENT',
+    APPROVED: 'APPROVED',
+    CARD_MISSING_ADDRESS: 'CARDMISSINGADDRESS',
+    CARD_ISSUED: 'CARDISSUED',
+    CARD_ISSUED_VIRTUAL: 'CARDISSUEDVIRTUAL',
+    CARD_REPLACED_VIRTUAL: 'CARDREPLACEDVIRTUAL',
+    CARD_REPLACED: 'CARDREPLACED',
+    CARD_ASSIGNED: 'CARDASSIGNED',
+    CARD_FROZEN: 'CARDFROZEN',
+    CARD_UNFROZEN: 'CARDUNFROZEN',
+    CARD_DEACTIVATED: 'CARDDEACTIVATED',
+    PERSONAL_CARD_CONNECTION_BROKEN: 'PERSONALCARDCONNECTIONBROKEN',
+    CHANGE_FIELD: 'CHANGEFIELD', // OldDot Action
+    CHANGE_POLICY: 'CHANGEPOLICY',
+    CREATED_REPORT_FOR_UNAPPROVED_TRANSACTIONS: 'CREATEDREPORTFORUNAPPROVEDTRANSACTIONS',
+    CHANGE_TYPE: 'CHANGETYPE', // OldDot Action
+    CHRONOS_OOO_LIST: 'CHRONOSOOOLIST',
+    CLOSED: 'CLOSED',
+    CREATED: 'CREATED',
+    DELETED_ACCOUNT: 'DELETEDACCOUNT', // Deprecated OldDot Action
+    DELETED_TRANSACTION: 'DELETEDTRANSACTION',
+    DEW_SUBMIT_FAILED: 'DEWSUBMITFAILED',
+    DEW_APPROVE_FAILED: 'DEWAPPROVEFAILED',
+    DISMISSED_VIOLATION: 'DISMISSEDVIOLATION',
+    DONATION: 'DONATION', // Deprecated OldDot Action
+    DYNAMIC_EXTERNAL_WORKFLOW_ROUTED: 'DYNAMICEXTERNALWORKFLOWROUTED',
+    EXPENSIFY_CARD_SYSTEM_MESSAGE: 'EXPENSIFYCARDSYSTEMMESSAGE',
+    EXPORTED_TO_CSV: 'EXPORTCSV', // OldDot Action
+    EXPORTED_TO_INTEGRATION: 'EXPORTINTEGRATION', // OldDot Action
+    EXPORTED_TO_QUICK_BOOKS: 'EXPORTED', // Deprecated OldDot Action
+    FIX_VIOLATION: 'FIXVIOLATION',
+    FORWARDED: 'FORWARDED', // OldDot Action
+    HOLD: 'HOLD',
+    HOLD_COMMENT: 'HOLDCOMMENT',
+    INTEGRATION_SYNC_FAILED: 'INTEGRATIONSYNCFAILED',
+    COMPANY_CARD_CONNECTION_BROKEN: 'COMPANYCARDCONNECTIONBROKEN',
+    COMMUTER_EXCLUSION: 'COMMUTEREXCLUSION',
+    PLAID_BALANCE_FAILURE: 'PLAIDBALANCEFAILURE',
+    IOU: 'IOU',
+    INTEGRATIONS_MESSAGE: 'INTEGRATIONSMESSAGE', // OldDot Action
+    MANAGER_ATTACH_RECEIPT: 'MANAGERATTACHRECEIPT', // OldDot Action
+    MANAGER_DETACH_RECEIPT: 'MANAGERDETACHRECEIPT', // OldDot Action
+    MARKED_REIMBURSED: 'MARKEDREIMBURSED', // OldDot Action
+    MARK_REIMBURSED_FROM_INTEGRATION: 'ACTIONMARKEDREIMBURSEDFROMINTEGRATION', // OldDot Action
+    MERGED_WITH_CASH_TRANSACTION: 'MERGEDWITHCASHTRANSACTION',
+    MODIFIED_EXPENSE: 'MODIFIEDEXPENSE',
+    CONCIERGE_AUTO_MATCH_VENDOR: 'CONCIERGEAUTOMATCHVENDOR',
+    MOVED: 'MOVED',
+    MOVED_TRANSACTION: 'MOVEDTRANSACTION',
+    UNREPORTED_TRANSACTION: 'UNREPORTEDTRANSACTION',
+    OUTDATED_BANK_ACCOUNT: 'OUTDATEDBANKACCOUNT', // OldDot Action
+    REIMBURSED: 'REIMBURSED',
+    REIMBURSEMENT_ACH_BOUNCE: 'REIMBURSEMENTACHBOUNCE', // OldDot Action
+    REIMBURSEMENT_ACH_CANCELED: 'REIMBURSEMENTACHCANCELED', // OldDot Action
+    REIMBURSEMENT_ACCOUNT_CHANGED: 'REIMBURSEMENTACCOUNTCHANGED', // OldDot Action
+    REIMBURSEMENT_DELAYED: 'REIMBURSEMENTDELAYED', // OldDot Action
+    REIMBURSEMENT_QUEUED: 'REIMBURSEMENTQUEUED',
+    REIMBURSEMENT_DEQUEUED: 'REIMBURSEMENTDEQUEUED',
+    REIMBURSEMENT_REQUESTED: 'REIMBURSEMENTREQUESTED', // Deprecated OldDot Action
+    REIMBURSEMENT_SETUP: 'REIMBURSEMENTSETUP', // Deprecated OldDot Action
+    REIMBURSEMENT_SETUP_REQUESTED: 'REIMBURSEMENTSETUPREQUESTED', // Deprecated OldDot Action
+    REIMBURSEMENT_DIRECTOR_INFORMATION_REQUIRED: 'DIRECTORINFORMATIONREQUIRED',
+    REJECTED: 'REJECTED',
+    REJECTED_TO_SUBMITTER: 'REJECTEDTOSUBMITTER',
+    REMOVED_FROM_APPROVAL_CHAIN: 'REMOVEDFROMAPPROVALCHAIN',
+    DEMOTED_FROM_WORKSPACE: 'DEMOTEDFROMWORKSPACE',
+    RENAMED: 'RENAMED',
+    RETRACTED: 'RETRACTED',
+    REOPENED: 'REOPENED',
+    REPORT_PREVIEW: 'REPORTPREVIEW',
+    REASSIGN_APPROVER: 'REASSIGNAPPROVER',
+    REROUTE: 'REROUTE',
+    SELECTED_FOR_RANDOM_AUDIT: 'SELECTEDFORRANDOMAUDIT', // OldDot Action
+    SETTLEMENT_ACCOUNT_LOCKED: 'SETTLEMENTACCOUNTLOCKED',
+    SHARE: 'SHARE', // OldDot Action
+    STRIPE_PAID: 'STRIPEPAID', // OldDot Action
+    SUBMITTED: 'SUBMITTED',
+    SUBMITTED_AND_CLOSED: 'SUBMITTEDCLOSED',
+    ACTION_DELEGATE_SUBMIT: 'DELEGATESUBMIT',
+    TAKE_CONTROL: 'TAKECONTROL', // OldDot Action
+    TASK_CANCELLED: 'TASKCANCELLED',
+    TASK_COMPLETED: 'TASKCOMPLETED',
+    TASK_EDITED: 'TASKEDITED',
+    TASK_REOPENED: 'TASKREOPENED',
+    TRAVEL_NUDGE: 'TRAVELNUDGE',
+    TRAVEL_UPDATE: 'TRAVEL_TRIP_ROOM_UPDATE',
+    TRIP_PREVIEW: 'TRIPPREVIEW',
+    UNAPPROVED: 'UNAPPROVED',
+    UNHOLD: 'UNHOLD',
+    UNSHARE: 'UNSHARE', // OldDot Action
+    UPDATE_GROUP_CHAT_MEMBER_ROLE: 'UPDATEGROUPCHATMEMBERROLE',
+    CONCIERGE_CATEGORY_OPTIONS: 'CONCIERGECATEGORYOPTIONS',
+    CONCIERGE_DESCRIPTION_OPTIONS: 'CONCIERGEDESCRIPTIONOPTIONS',
+    CONCIERGE_AUTO_MAP_MCC_GROUPS: 'CONCIERGEAUTOMAPMCCGROUPS',
+    POLICY_CHANGE_LOG: {
+        ADD_APPROVER_RULE: 'POLICYCHANGELOG_ADD_APPROVER_RULE',
+        ADD_BUDGET: 'POLICYCHANGELOG_ADD_BUDGET',
+        ADD_CATEGORY: 'POLICYCHANGELOG_ADD_CATEGORY',
+        ADD_CUSTOM_UNIT: 'POLICYCHANGELOG_ADD_CUSTOM_UNIT',
+        ADD_CUSTOM_UNIT_RATE: 'POLICYCHANGELOG_ADD_CUSTOM_UNIT_RATE',
+        ADD_EMPLOYEE: 'POLICYCHANGELOG_ADD_EMPLOYEE',
+        ADD_CARD_FEED: 'POLICYCHANGELOG_ADD_CARD_FEED',
+        ADD_EXPENSIFY_CARD_RULE: 'POLICYCHANGELOG_ADD_EXPENSIFY_CARD_RULE',
+        ADD_AGENT_RULE: 'POLICYCHANGELOG_ADD_AGENT_RULE',
+        UPDATE_AGENT_RULE: 'POLICYCHANGELOG_UPDATE_AGENT_RULE',
+        DELETE_AGENT_RULE: 'POLICYCHANGELOG_DELETE_AGENT_RULE',
+        ADD_RULE: 'POLICYCHANGELOG_ADD_RULE',
+        UPDATE_RULE: 'POLICYCHANGELOG_UPDATE_RULE',
+        REMOVE_RULE: 'POLICYCHANGELOG_REMOVE_RULE',
+        ADD_INTEGRATION: 'POLICYCHANGELOG_ADD_INTEGRATION',
+        ADD_REPORT_FIELD: 'POLICYCHANGELOG_ADD_REPORT_FIELD',
+        ADD_TAG: 'POLICYCHANGELOG_ADD_TAG',
+        ADD_TAX: 'POLICYCHANGELOG_ADD_TAX',
+        DELETE_TAX: 'POLICYCHANGELOG_DELETE_TAX',
+        UPDATE_TAX: 'POLICYCHANGELOG_UPDATE_TAX',
+        DELETE_ALL_TAGS: 'POLICYCHANGELOG_DELETE_ALL_TAGS',
+        DELETE_APPROVER_RULE: 'POLICYCHANGELOG_DELETE_APPROVER_RULE',
+        DELETE_BUDGET: 'POLICYCHANGELOG_DELETE_BUDGET',
+        DELETE_CATEGORY: 'POLICYCHANGELOG_DELETE_CATEGORY',
+        DELETE_CATEGORIES: 'POLICYCHANGELOG_DELETE_CATEGORIES',
+        DELETE_CUSTOM_UNIT: 'POLICYCHANGELOG_DELETE_CUSTOM_UNIT',
+        DELETE_CUSTOM_UNIT_RATE: 'POLICYCHANGELOG_DELETE_CUSTOM_UNIT_RATE',
+        DELETE_CUSTOM_UNIT_SUB_RATE: 'POLICYCHANGELOG_DELETE_CUSTOM_UNIT_SUB_RATE',
+        DELETE_EMPLOYEE: 'POLICYCHANGELOG_DELETE_EMPLOYEE',
+        DELETE_CARD_FEED: 'POLICYCHANGELOG_DELETE_CARD_FEED',
+        DELETE_INTEGRATION: 'POLICYCHANGELOG_DELETE_INTEGRATION',
+        DELETE_REPORT_FIELD: 'POLICYCHANGELOG_DELETE_REPORT_FIELD',
+        DELETE_TAG: 'POLICYCHANGELOG_DELETE_TAG',
+        DELETE_MULTIPLE_TAGS: 'POLICYCHANGELOG_DELETE_MULTIPLE_TAGS',
+        IMPORT_CUSTOM_UNIT_RATES: 'POLICYCHANGELOG_IMPORT_CUSTOM_UNIT_RATES',
+        IMPORT_TAGS: 'POLICYCHANGELOG_IMPORT_TAGS',
+        INDIVIDUAL_BUDGET_NOTIFICATION: 'POLICYCHANGELOG_INDIVIDUAL_BUDGET_NOTIFICATION',
+        INVITE_TO_ROOM: 'POLICYCHANGELOG_INVITETOROOM',
+        REMOVE_FROM_ROOM: 'POLICYCHANGELOG_REMOVEFROMROOM',
+        REMOVE_EXPENSIFY_CARD_RULE: 'POLICYCHANGELOG_REMOVE_EXPENSIFY_CARD_RULE',
+        LEAVE_ROOM: 'POLICYCHANGELOG_LEAVEROOM',
+        REPLACE_CATEGORIES: 'POLICYCHANGELOG_REPLACE_CATEGORIES',
+        SET_AUTO_REIMBURSEMENT: 'POLICYCHANGELOG_SET_AUTOREIMBURSEMENT',
+        SET_AUTO_JOIN: 'POLICYCHANGELOG_SET_AUTO_JOIN',
+        SET_CATEGORY_NAME: 'POLICYCHANGELOG_SET_CATEGORY_NAME',
+        SHARED_BUDGET_NOTIFICATION: 'POLICYCHANGELOG_SHARED_BUDGET_NOTIFICATION',
+        UPDATE_ACH_ACCOUNT: 'POLICYCHANGELOG_UPDATE_ACH_ACCOUNT',
+        UPDATE_APPROVER_RULE: 'POLICYCHANGELOG_UPDATE_APPROVER_RULE',
+        UPDATE_AUDIT_RATE: 'POLICYCHANGELOG_UPDATE_AUDIT_RATE',
+        UPDATE_AUTO_HARVESTING: 'POLICYCHANGELOG_UPDATE_AUTOHARVESTING',
+        UPDATE_AUTO_REIMBURSEMENT: 'POLICYCHANGELOG_UPDATE_AUTOREIMBURSEMENT',
+        UPDATE_AUTO_PAY_APPROVED_REPORTS_ENABLED: 'POLICYCHANGELOG_UPDATE_AUTO_PAY_APPROVED_REPORTS_ENABLED',
+        UPDATE_MCC_GROUP_CATEGORY: 'POLICYCHANGELOG_UPDATE_MCC_GROUP_CATEGORY',
+        UPDATE_AUTO_REPORTING_FREQUENCY: 'POLICYCHANGELOG_UPDATE_AUTOREPORTING_FREQUENCY',
+        UPDATE_BUDGET: 'POLICYCHANGELOG_UPDATE_BUDGET',
+        UPDATE_CATEGORY: 'POLICYCHANGELOG_UPDATE_CATEGORY',
+        UPDATE_CATEGORY_TAX_RATE: 'POLICYCHANGELOG_UPDATE_CATEGORY_TAX_RATE',
+        UPDATE_CATEGORIES: 'POLICYCHANGELOG_UPDATE_CATEGORIES',
+        UPDATE_CURRENCY: 'POLICYCHANGELOG_UPDATE_CURRENCY',
+        UPDATE_CUSTOM_UNIT: 'POLICYCHANGELOG_UPDATE_CUSTOM_UNIT',
+        UPDATE_CUSTOM_UNIT_RATE: 'POLICYCHANGELOG_UPDATE_CUSTOM_UNIT_RATE',
+        UPDATE_CUSTOM_UNIT_SUB_RATE: 'POLICYCHANGELOG_UPDATE_CUSTOM_UNIT_SUB_RATE',
+        UPDATE_DEFAULT_BILLABLE: 'POLICYCHANGELOG_UPDATE_DEFAULT_BILLABLE',
+        UPDATE_DEFAULT_REIMBURSABLE: 'POLICYCHANGELOG_UPDATE_DEFAULT_REIMBURSABLE',
+        UPDATE_DEFAULT_TITLE: 'POLICYCHANGELOG_UPDATE_DEFAULT_TITLE',
+        UPDATE_DEFAULT_TITLE_ENFORCED: 'POLICYCHANGELOG_UPDATE_DEFAULT_TITLE_ENFORCED',
+        UPDATE_DISABLED_FIELDS: 'POLICYCHANGELOG_UPDATE_DISABLED_FIELDS',
+        UPDATE_EMPLOYEE: 'POLICYCHANGELOG_UPDATE_EMPLOYEE',
+        UPDATE_EXPENSIFY_CARD_RULE: 'POLICYCHANGELOG_UPDATE_EXPENSIFY_CARD_RULE',
+        UPDATE_FIELD: 'POLICYCHANGELOG_UPDATE_FIELD',
+        UPDATE_ADDRESS: 'POLICYCHANGELOG_UPDATE_ADDRESS',
+        UPDATE_FEATURE_ENABLED: 'POLICYCHANGELOG_UPDATE_FEATURE_ENABLED',
+        UPDATE_IS_ATTENDEE_TRACKING_ENABLED: 'POLICYCHANGELOG_UPDATE_IS_ATTENDEE_TRACKING_ENABLED',
+        UPDATE_REQUIRE_COMPANY_CARDS_ENABLED: 'POLICYCHANGELOG_UPDATE_REQUIRE_COMPANY_CARDS_ENABLED',
+        UPDATE_REQUIRES_CATEGORY: 'POLICYCHANGELOG_UPDATE_REQUIRES_CATEGORY',
+        UPDATE_REQUIRES_TAG: 'POLICYCHANGELOG_UPDATE_REQUIRES_TAG',
+        UPDATE_GLOBAL_REIMBURSEMENTS_FX_PREFERENCE: 'POLICYCHANGELOG_UPDATE_GLOBAL_REIMBURSEMENTS_FX_PREFERENCE',
+        UPDATE_DEFAULT_APPROVER: 'POLICYCHANGELOG_UPDATE_DEFAULT_APPROVER',
+        UPDATE_SUBMITS_TO: 'POLICYCHANGELOG_UPDATE_SUBMITS_TO',
+        UPDATE_FORWARDS_TO: 'POLICYCHANGELOG_UPDATE_FORWARDS_TO',
+        UPDATE_CUSTOM_TAX_NAME: 'POLICYCHANGELOG_UPDATE_CUSTOM_TAX_NAME',
+        UPDATE_CURRENCY_DEFAULT_TAX: 'POLICYCHANGELOG_UPDATE_CURRENCY_DEFAULT_TAX',
+        UPDATE_FOREIGN_CURRENCY_DEFAULT_TAX: 'POLICYCHANGELOG_UPDATE_FOREIGN_CURRENCY_DEFAULT_TAX',
+        UPDATE_INVOICE_COMPANY_NAME: 'POLICYCHANGELOG_UPDATE_INVOICE_COMPANY_NAME',
+        UPDATE_INVOICE_COMPANY_WEBSITE: 'POLICYCHANGELOG_UPDATE_INVOICE_COMPANY_WEBSITE',
+        UPDATE_MANUAL_APPROVAL_THRESHOLD: 'POLICYCHANGELOG_UPDATE_MANUAL_APPROVAL_THRESHOLD',
+        UPDATE_MAX_EXPENSE_AMOUNT: 'POLICYCHANGELOG_UPDATE_MAX_EXPENSE_AMOUNT',
+        UPDATE_MAX_EXPENSE_AMOUNT_NO_RECEIPT: 'POLICYCHANGELOG_UPDATE_MAX_EXPENSE_AMOUNT_NO_RECEIPT',
+        UPDATE_MAX_EXPENSE_AMOUNT_NO_ITEMIZED_RECEIPT: 'POLICYCHANGELOG_UPDATE_MAX_EXPENSE_AMOUNT_NO_ITEMIZED_RECEIPT',
+        UPDATE_MAX_EXPENSE_AGE: 'POLICYCHANGELOG_UPDATE_MAX_EXPENSE_AGE',
+        UPDATE_MULTIPLE_TAGS_APPROVER_RULES: 'POLICYCHANGELOG_UPDATE_MULTIPLE_TAGS_APPROVER_RULES',
+        UPDATE_NAME: 'POLICYCHANGELOG_UPDATE_NAME',
+        UPDATE_DESCRIPTION: 'POLICYCHANGELOG_UPDATE_DESCRIPTION',
+        UPDATE_OWNERSHIP: 'POLICYCHANGELOG_UPDATE_OWNERSHIP',
+        UPDATE_REIMBURSER: 'POLICYCHANGELOG_UPDATE_REIMBURSER',
+        UPDATE_PROHIBITED_EXPENSES: 'POLICYCHANGELOG_UPDATE_PROHIBITED_EXPENSES',
+        UPDATE_COMMUTER_EXCLUSIONS: 'POLICYCHANGELOG_UPDATE_COMMUTER_EXCLUSIONS',
+        UPDATE_REIMBURSEMENT_CHOICE: 'POLICYCHANGELOG_UPDATE_REIMBURSEMENT_CHOICE',
+        UPDATE_REIMBURSEMENT_ENABLED: 'POLICYCHANGELOG_UPDATE_REIMBURSEMENT_ENABLED',
+        UPDATE_REPORT_FIELD: 'POLICYCHANGELOG_UPDATE_REPORT_FIELD',
+        UPDATE_TAG: 'POLICYCHANGELOG_UPDATE_TAG',
+        UPDATE_TAG_ENABLED: 'POLICYCHANGELOG_UPDATE_TAG_ENABLED',
+        UPDATE_TAG_LIST: 'POLICYCHANGELOG_UPDATE_TAG_LIST',
+        UPDATE_TAG_LIST_NAME: 'POLICYCHANGELOG_UPDATE_TAG_LIST_NAME',
+        UPDATE_TAG_LIST_REQUIRED: 'POLICYCHANGELOG_UPDATE_TAG_LIST_REQUIRED',
+        UPDATE_TAG_NAME: 'POLICYCHANGELOG_UPDATE_TAG_NAME',
+        UPDATE_TIME_ENABLED: 'POLICYCHANGELOG_UPDATE_TIME_ENABLED',
+        UPDATE_TIME_RATE: 'POLICYCHANGELOG_UPDATE_TIME_RATE',
+        RENAME_CARD_FEED: 'POLICYCHANGELOG_RENAME_CARD_FEED',
+        ASSIGN_COMPANY_CARD: 'POLICYCHANGELOG_ASSIGN_COMPANY_CARD',
+        UNASSIGN_COMPANY_CARD: 'POLICYCHANGELOG_UNASSIGN_COMPANY_CARD',
+        UPDATE_CARD_FEED_LIABILITY: 'POLICYCHANGELOG_UPDATE_CARD_FEED_LIABILITY',
+        UPDATE_CARD_FEED_STATEMENT_PERIOD: 'POLICYCHANGELOG_UPDATE_CARD_FEED_STATEMENT_PERIOD',
+        LEAVE_POLICY: 'POLICYCHANGELOG_LEAVE_POLICY',
+        CORPORATE_UPGRADE: 'POLICYCHANGELOG_CORPORATE_UPGRADE',
+        CORPORATE_FORCE_UPGRADE: 'POLICYCHANGELOG_CORPORATE_FORCE_UPGRADE',
+        TEAM_DOWNGRADE: 'POLICYCHANGELOG_TEAM_DOWNGRADE',
+        COPY_OVERVIEW: 'POLICYCHANGELOG_COPY_OVERVIEW',
+        COPY_CURRENCY: 'POLICYCHANGELOG_COPY_CURRENCY',
+        COPY_EMPLOYEES: 'POLICYCHANGELOG_COPY_EMPLOYEES',
+        COPY_REPORT_FIELDS: 'POLICYCHANGELOG_COPY_REPORT_FIELDS',
+        COPY_ACCOUNTING: 'POLICYCHANGELOG_COPY_ACCOUNTING',
+        COPY_RECEIPT_PARTNERS: 'POLICYCHANGELOG_COPY_RECEIPT_PARTNERS',
+        COPY_HR: 'POLICYCHANGELOG_COPY_HR',
+        COPY_CATEGORIES: 'POLICYCHANGELOG_COPY_CATEGORIES',
+        COPY_TAGS: 'POLICYCHANGELOG_COPY_TAGS',
+        COPY_TAXES: 'POLICYCHANGELOG_COPY_TAXES',
+        COPY_TIME_TRACKING: 'POLICYCHANGELOG_COPY_TIME_TRACKING',
+        COPY_WORKFLOWS: 'POLICYCHANGELOG_COPY_WORKFLOWS',
+        COPY_RULES: 'POLICYCHANGELOG_COPY_RULES',
+        COPY_CODING_RULES: 'POLICYCHANGELOG_COPY_CODING_RULES',
+        COPY_DISTANCE: 'POLICYCHANGELOG_COPY_DISTANCE',
+        COPY_PER_DIEM: 'POLICYCHANGELOG_COPY_PER_DIEM',
+        COPY_INVOICES: 'POLICYCHANGELOG_COPY_INVOICES',
+        COPY_TRAVEL: 'POLICYCHANGELOG_COPY_TRAVEL',
+    },
+    RECEIPT_SCAN_FAILED: 'RECEIPTSCANFAILED',
+    RESOLVED_DUPLICATES: 'RESOLVEDDUPLICATES',
+    ROOM_CHANGE_LOG: {
+        INVITE_TO_ROOM: 'INVITETOROOM',
+        REMOVE_FROM_ROOM: 'REMOVEFROMROOM',
+        LEAVE_ROOM: 'LEAVEROOM',
+        UPDATE_ROOM_DESCRIPTION: 'UPDATEROOMDESCRIPTION',
+        UPDATE_ROOM_AVATAR: 'UPDATEROOMAVATAR',
+    },
+    REJECTEDTRANSACTION_THREAD: 'REJECTEDTRANSACTION_THREAD',
+    REJECTED_TRANSACTION_MARKASRESOLVED: 'REJECTEDTRANSACTIONMARKASRESOLVED',
+} as const;
+// cspell:enable
+
+// eslint-disable-next-line import/prefer-default-export -- Preserve the named API used by the extracted CONST module.
+export {REPORT_ACTION_TYPES};
