+++ src/CONST/language/Policy.ts
@@ -1,0 +1,160 @@
+const POLICY = {
+    ROLE: {
+        OWNER: 'owner',
+        ADMIN: 'admin',
+        AUDITOR: 'auditor',
+        USER: 'user',
+        EDITOR: 'editor',
+        CARD_ADMIN: 'cardAdmin',
+        PEOPLE_ADMIN: 'peopleAdmin',
+        PAYMENTS_ADMIN: 'paymentsAdmin',
+    },
+    REIMBURSEMENT_CHOICES: {
+        REIMBURSEMENT_YES: 'reimburseYes',
+        REIMBURSEMENT_NO: 'reimburseNo',
+        REIMBURSEMENT_MANUAL: 'reimburseManual',
+    },
+    RECEIPT_PARTNERS: {
+        UBER_EMPLOYEE_STATUS: {
+            CREATED: 'CREATED',
+            INVITED: 'INVITED',
+            LINKED_PENDING_APPROVAL: 'LINKED_PENDING_APPROVAL',
+            LINKED: 'LINKED',
+            SUSPENDED: 'SUSPENDED',
+            DELETED: 'DELETED',
+            NONE: 'NONE',
+        },
+    },
+    CONNECTIONS: {
+        ACCOUNTING_INTEGRATION_ALIASES: {
+            INTUIT_ENTERPRISE_SUITE: 'intuitEnterpriseSuite',
+        },
+        NAME: {
+            QBO: 'quickbooksOnline',
+            QBD: 'quickbooksDesktop',
+            XERO: 'xero',
+            NETSUITE: 'netsuite',
+            SAGE_INTACCT: 'intacct',
+            CERTINIA: 'financialforce',
+            RILLET: 'rillet',
+            DUALENTRY: 'dualEntry',
+            GUSTO: 'gusto',
+            ZENEFITS: 'zenefits',
+            MERGE_HR: 'merge_hris',
+        },
+        NAME_USER_FRIENDLY: {
+            netsuite: 'NetSuite',
+            quickbooksOnline: 'QuickBooks Online',
+            quickbooksDesktop: 'QuickBooks Desktop',
+            xero: 'Xero',
+            intacct: 'Sage Intacct',
+            financialforce: 'Certinia',
+            rillet: 'Rillet',
+            dualEntry: 'DualEntry',
+            gusto: 'Gusto',
+            billCom: 'Bill.com',
+            zenefits: 'TriNet',
+            // eslint-disable-next-line @typescript-eslint/naming-convention -- Matches the integration slug returned by the API.
+            merge_hris: 'Merge HR',
+            sap: 'SAP',
+            oracle: 'Oracle',
+            microsoftDynamics: 'Microsoft Dynamics',
+            other: 'Other',
+        },
+        SYNC_STAGE_NAME: {
+            STARTING_IMPORT_QBO: 'startingImportQBO',
+            STARTING_IMPORT_XERO: 'startingImportXero',
+            STARTING_IMPORT_QBD: 'startingImportQBD',
+            QBO_IMPORT_MAIN: 'quickbooksOnlineImportMain',
+            QBO_IMPORT_CUSTOMERS: 'quickbooksOnlineImportCustomers',
+            QBO_IMPORT_EMPLOYEES: 'quickbooksOnlineImportEmployees',
+            QBO_IMPORT_ACCOUNTS: 'quickbooksOnlineImportAccounts',
+            QBO_IMPORT_CLASSES: 'quickbooksOnlineImportClasses',
+            QBO_IMPORT_LOCATIONS: 'quickbooksOnlineImportLocations',
+            QBO_IMPORT_PROCESSING: 'quickbooksOnlineImportProcessing',
+            QBO_SYNC_PAYMENTS: 'quickbooksOnlineSyncBillPayments',
+            QBO_IMPORT_TAX_CODES: 'quickbooksOnlineSyncTaxCodes',
+            QBO_CHECK_CONNECTION: 'quickbooksOnlineCheckConnection',
+            QBO_SYNC_TITLE: 'quickbooksOnlineSyncTitle',
+            QBO_SYNC_LOAD_DATA: 'quickbooksOnlineSyncLoadData',
+            QBO_SYNC_APPLY_CATEGORIES: 'quickbooksOnlineSyncApplyCategories',
+            QBO_SYNC_APPLY_CUSTOMERS: 'quickbooksOnlineSyncApplyCustomers',
+            QBO_SYNC_APPLY_PEOPLE: 'quickbooksOnlineSyncApplyEmployees',
+            QBO_SYNC_APPLY_CLASSES_LOCATIONS: 'quickbooksOnlineSyncApplyClassesLocations',
+            QBD_IMPORT_TITLE: 'quickbooksDesktopImportTitle',
+            QBD_IMPORT_ACCOUNTS: 'quickbooksDesktopImportAccounts',
+            QBD_IMPORT_APPROVE_CERTIFICATE: 'quickbooksDesktopImportApproveCertificate',
+            QBD_IMPORT_DIMENSIONS: 'quickbooksDesktopImportDimensions',
+            QBD_IMPORT_CLASSES: 'quickbooksDesktopImportClasses',
+            QBD_IMPORT_CUSTOMERS: 'quickbooksDesktopImportCustomers',
+            QBD_IMPORT_VENDORS: 'quickbooksDesktopImportVendors',
+            QBD_IMPORT_EMPLOYEES: 'quickbooksDesktopImportEmployees',
+            QBD_IMPORT_MORE: 'quickbooksDesktopImportMore',
+            QBD_IMPORT_GENERIC: 'quickbooksDesktopImportSavePolicy',
+            QBD_WEB_CONNECTOR_REMINDER: 'quickbooksDesktopWebConnectorReminder',
+            JOB_DONE: 'jobDone',
+            XERO_SYNC_STEP: 'xeroSyncStep',
+            XERO_SYNC_XERO_REIMBURSED_REPORTS: 'xeroSyncXeroReimbursedReports',
+            XERO_SYNC_EXPENSIFY_REIMBURSED_REPORTS: 'xeroSyncExpensifyReimbursedReports',
+            XERO_SYNC_IMPORT_CHART_OF_ACCOUNTS: 'xeroSyncImportChartOfAccounts',
+            XERO_SYNC_IMPORT_CATEGORIES: 'xeroSyncImportCategories',
+            XERO_SYNC_IMPORT_TRACKING_CATEGORIES: 'xeroSyncImportTrackingCategories',
+            XERO_SYNC_IMPORT_CUSTOMERS: 'xeroSyncImportCustomers',
+            XERO_SYNC_IMPORT_BANK_ACCOUNTS: 'xeroSyncImportBankAccounts',
+            XERO_SYNC_IMPORT_TAX_RATES: 'xeroSyncImportTaxRates',
+            XERO_CHECK_CONNECTION: 'xeroCheckConnection',
+            XERO_SYNC_TITLE: 'xeroSyncTitle',
+            NETSUITE_SYNC_CONNECTION: 'netSuiteSyncConnection',
+            NETSUITE_SYNC_CUSTOMERS: 'netSuiteSyncCustomers',
+            NETSUITE_SYNC_INIT_DATA: 'netSuiteSyncInitData',
+            NETSUITE_SYNC_IMPORT_TAXES: 'netSuiteSyncImportTaxes',
+            NETSUITE_SYNC_IMPORT_ITEMS: 'netSuiteSyncImportItems',
+            NETSUITE_SYNC_DATA: 'netSuiteSyncData',
+            NETSUITE_SYNC_ACCOUNTS: 'netSuiteSyncAccounts',
+            NETSUITE_SYNC_CURRENCIES: 'netSuiteSyncCurrencies',
+            NETSUITE_SYNC_CATEGORIES: 'netSuiteSyncCategories',
+            NETSUITE_SYNC_IMPORT_CUSTOM_LISTS: 'netSuiteSyncImportCustomLists',
+            NETSUITE_SYNC_IMPORT_EMPLOYEES: 'netSuiteSyncImportEmployees',
+            NETSUITE_SYNC_IMPORT_SUBSIDIARIES: 'netSuiteSyncImportSubsidiaries',
+            NETSUITE_SYNC_IMPORT_VENDORS: 'netSuiteSyncImportVendors',
+            NETSUITE_SYNC_REPORT_FIELDS: 'netSuiteSyncReportFields',
+            NETSUITE_SYNC_TAGS: 'netSuiteSyncTags',
+            NETSUITE_SYNC_UPDATE_DATA: 'netSuiteSyncUpdateConnectionData',
+            NETSUITE_SYNC_NETSUITE_REIMBURSED_REPORTS: 'netSuiteSyncNetSuiteReimbursedReports',
+            NETSUITE_SYNC_EXPENSIFY_REIMBURSED_REPORTS: 'netSuiteSyncExpensifyReimbursedReports',
+            NETSUITE_SYNC_IMPORT_VENDORS_TITLE: 'netSuiteImportVendorsTitle',
+            NETSUITE_SYNC_IMPORT_CUSTOM_LISTS_TITLE: 'netSuiteImportCustomListsTitle',
+            SAGE_INTACCT_SYNC_CHECK_CONNECTION: 'intacctCheckConnection',
+            SAGE_INTACCT_SYNC_IMPORT_TITLE: 'intacctImportTitle',
+            SAGE_INTACCT_SYNC_IMPORT_DATA: 'intacctImportData',
+            SAGE_INTACCT_SYNC_IMPORT_EMPLOYEES: 'intacctImportEmployees',
+            SAGE_INTACCT_SYNC_IMPORT_DIMENSIONS: 'intacctImportDimensions',
+            SAGE_INTACCT_SYNC_IMPORT_SYNC_REIMBURSED_REPORTS: 'intacctImportSyncBillPayments',
+            GUSTO_SYNC_TITLE: 'gustoSyncTitle',
+            GUSTO_SYNC_LOAD_DATA: 'gustoSyncLoadData',
+            GUSTO_SYNC_PROVISIONING: 'gustoSyncProvisioning',
+            ZENEFITS_SYNC_TITLE: 'zenefitsSyncTitle',
+            ZENEFITS_SYNC_LOAD_DATA: 'zenefitsSyncLoadData',
+            ZENEFITS_SYNC_PROVISIONING: 'zenefitsSyncProvisioning',
+            FINANCIAL_FORCE_SYNC_TITLE: 'financialForceSyncTitle',
+            FINANCIAL_FORCE_SYNC_STEP: 'financialForceSyncStep',
+            FINANCIAL_FORCE_SYNC_CATEGORIES: 'financialForceSyncCategories',
+            FINANCIAL_FORCE_SYNC_TAGS: 'financialForceSyncTags',
+            FINANCIAL_FORCE_SYNC_VENDORS: 'financialForceSyncVendors',
+            FINANCIAL_FORCE_SYNC_CONTACTS: 'financialForceSyncContacts',
+            FINANCIAL_FORCE_SYNC_COMPANIES: 'financialForceSyncCompanies',
+            FINANCIAL_FORCE_SYNC_USERS: 'financialForceSyncUsers',
+            FINANCIAL_FORCE_SYNC_DIMENSIONS: 'financialForceSyncDimensions',
+            FINANCIAL_FORCE_SYNC_MARK_REIMBURSED: 'financialForceMarkAsReimbursed',
+            RILLET_SYNC_TITLE: 'rilletSyncTitle',
+            RILLET_SYNC_CONNECTION: 'rilletSyncConnection',
+            RILLET_SYNC_IMPORT_DATA: 'rilletSyncImportData',
+            DUALENTRY_SYNC_TITLE: 'dualEntrySyncTitle',
+            DUALENTRY_SYNC_CONNECTION: 'dualEntrySyncConnection',
+            DUALENTRY_SYNC_IMPORT_DATA: 'dualEntrySyncImportData',
+        },
+    },
+} as const;
+
+// eslint-disable-next-line import/prefer-default-export -- Preserve the named API used by the extracted CONST module.
+export {POLICY};
