+++ src/CONST/language/TransactionViolation.ts
@@ -1,0 +1,8 @@
+const MODIFIED_AMOUNT_VIOLATION_DATA = {
+    DISTANCE: 'distance',
+    CARD: 'card',
+    SMARTSCAN: 'smartscan',
+} as const;
+
+// eslint-disable-next-line import/prefer-default-export -- Preserve the named API used by the extracted CONST module.
+export {MODIFIED_AMOUNT_VIOLATION_DATA};
