+++ src/CONST/language/Wallet.ts
@@ -1,0 +1,9 @@
+const WALLET = {
+    PROGRAM_ISSUERS: {
+        EXPENSIFY_PAYMENTS: 'Expensify Payments LLC',
+        BANCORP_BANK: 'The Bancorp Bank, N.A.',
+    },
+} as const;
+
+// eslint-disable-next-line import/prefer-default-export -- Preserve the named API used by the extracted CONST module.
+export {WALLET};
