+++ src/CONST/language/index.ts
@@ -1,0 +1,198 @@
+import {COUNTRY, ALL_COUNTRIES} from './Countries';
+import {CERTINIA_EXPORT_DATE, CERTINIA_EXPORT_STATUS, CERTINIA_MAPPING_VALUE, CERTINIA_PARENT_TAG_MAPPING, CERTINIA_REPORT_EXPORT_STATUS} from './integrations/Certinia';
+import {DUALENTRY_EXPORT_DATE, DUALENTRY_EXPORT_NON_REIMBURSABLE, DUALENTRY_EXPORT_REIMBURSABLE} from './integrations/DualEntry';
+import {INTEGRATION_ENTITY_MAP_TYPES} from './integrations/EntityMapping';
+import {
+    NETSUITE_EXPORT_DATE,
+    NETSUITE_EXPORT_DESTINATION,
+    NETSUITE_IMPORT,
+    NETSUITE_INVOICE_ITEM_PREFERENCE,
+    NETSUITE_JOURNALS_APPROVAL_LEVEL,
+    NETSUITE_JOURNAL_POSTING_PREFERENCE,
+    NETSUITE_REPORTS_APPROVAL_LEVEL,
+    NETSUITE_VENDOR_BILLS_APPROVAL_LEVEL,
+} from './integrations/NetSuite';
+import {
+    QUICKBOOKS_DESKTOP_NON_REIMBURSABLE_EXPORT_ACCOUNT_TYPE,
+    QUICKBOOKS_DESKTOP_REIMBURSABLE_ACCOUNT_TYPE,
+    QUICKBOOKS_EXPORT_DATE,
+    QUICKBOOKS_NON_REIMBURSABLE_EXPORT_ACCOUNT_TYPE,
+    QUICKBOOKS_REIMBURSABLE_ACCOUNT_TYPE,
+} from './integrations/QuickBooks';
+import {RILLET_EXPORT_DATE, RILLET_EXPORT_NON_REIMBURSABLE, RILLET_EXPORT_REIMBURSABLE} from './integrations/Rillet';
+import {SAGE_INTACCT_CONFIG, SAGE_INTACCT_EXPORT_DATE, SAGE_INTACCT_NON_REIMBURSABLE_EXPENSE_TYPE, SAGE_INTACCT_REIMBURSABLE_EXPENSE_TYPE} from './integrations/SageIntacct';
+import {XERO_CONFIG, XERO_EXPORT_DATE} from './integrations/Xero';
+import {
+    COMPANY_CARDS,
+    CUSTOM_UNITS,
+    DELEGATE,
+    DELEGATE_ROLE,
+    DOT_SEPARATOR,
+    EXIT_SURVEY,
+    IMPORT_SPREADSHEET,
+    INTRO_CHOICES,
+    REFERRAL_PROGRAM,
+    REPORT_VIOLATIONS,
+    RTER_VIOLATION_TYPES,
+    UPGRADE_FEATURE_INTRO_MAPPING,
+} from './Misc';
+import {ONBOARDING_ACCOUNTING_MAPPING, ONBOARDING_CHOICES, ONBOARDING_COMPANY_SIZE, ONBOARDING_PERSONAL_TRACK_GOALS, ONBOARDING_TASK_TYPE} from './Onboarding';
+import {POLICY} from './Policy';
+import {IOU, NEXT_STEP, REPORT} from './Report';
+import {SEARCH} from './Search';
+import {SPEND_RULES} from './SpendRules';
+import {MODIFIED_AMOUNT_VIOLATION_DATA} from './TransactionViolation';
+import {
+    AI_FEATURES_PROMO_LEARN_MORE_URLS,
+    APP_DOWNLOAD_LINKS,
+    CFPB_PREPAID_URL,
+    COLLECT_UPGRADE_HELP_URL,
+    COMPANY_CARDS_AMEX_COMMERCIAL_CARD_HELP,
+    COMPANY_CARDS_CONNECT_CREDIT_CARDS_HELP_URL,
+    COMPANY_CARDS_CREATE_FILE_FEED_HELP_URL,
+    COMPANY_CARDS_MASTERCARD_COMMERCIAL_CARDS,
+    COMPANY_CARDS_STRIPE_HELP,
+    COMPANY_CARDS_VISA_COMMERCIAL_CARD_HELP,
+    CONCIERGE_CHAT_NAME,
+    CONCIERGE_EXPLAIN_LINK_PATH,
+    CONFIGURE_EXPENSE_REPORT_RULES_HELP_URL,
+    CUSTOM_AGENTS_HELP_URL,
+    CUSTOM_REPORT_NAME_HELP_URL,
+    DEEP_DIVE_ERECEIPTS,
+    DEEP_DIVE_EXPENSIFY_CARD,
+    DEEP_DIVE_PER_DIEM,
+    DELEGATE_ROLE_HELP_DOT_ARTICLE_LINK,
+    DENIED_CAMERA_ACCESS_INSTRUCTIONS_URL,
+    DOMAIN_VERIFICATION_HELP_URL,
+    ELECTRONIC_DISCLOSURES_URL,
+    EMAIL,
+    ENABLE_GLOBAL_REIMBURSEMENT_HELP_URL,
+    EXAMPLE_PHONE_NUMBER,
+    FEES_URL,
+    FORMATTED_EXAMPLE_PHONE_NUMBER,
+    HELP_DOC_LINKS,
+    IMPORT_TAGS_EXPENSIFY_URL,
+    IMPORT_TAGS_EXPENSIFY_URL_DEPENDENT_TAGS,
+    NEW_EXPENSIFY_URL,
+    OLD_DOT_PUBLIC_URLS,
+    ONFIDO_FACIAL_SCAN_POLICY_URL,
+    ONFIDO_PRIVACY_POLICY_URL,
+    ONFIDO_TERMS_OF_SERVICE_URL,
+    PERSONAL_DATA_PROTECTION_INFO_URL,
+    PRICING,
+    SAML_HELP_URL,
+    SELECT_WORKFLOWS_HELP_URL,
+    SET_NOTIFICATION_LINK,
+    TERMS,
+    TRAVEL_TERMS_URL,
+    UBER_TERMS_LINK,
+    UNLOCK_BANK_ACCOUNT_HELP_URL,
+} from './URLs';
+import {WALLET} from './Wallet';
+
+const CONST = {
+    AI_FEATURES_PROMO_LEARN_MORE_URLS,
+    APP_DOWNLOAD_LINKS,
+    CFPB_PREPAID_URL,
+    COLLECT_UPGRADE_HELP_URL,
+    COMPANY_CARDS_AMEX_COMMERCIAL_CARD_HELP,
+    COMPANY_CARDS_CONNECT_CREDIT_CARDS_HELP_URL,
+    COMPANY_CARDS_CREATE_FILE_FEED_HELP_URL,
+    COMPANY_CARDS_MASTERCARD_COMMERCIAL_CARDS,
+    COMPANY_CARDS_STRIPE_HELP,
+    COMPANY_CARDS_VISA_COMMERCIAL_CARD_HELP,
+    CONCIERGE_CHAT_NAME,
+    CONCIERGE_EXPLAIN_LINK_PATH,
+    CONFIGURE_EXPENSE_REPORT_RULES_HELP_URL,
+    CUSTOM_AGENTS_HELP_URL,
+    CUSTOM_REPORT_NAME_HELP_URL,
+    DEEP_DIVE_ERECEIPTS,
+    DEEP_DIVE_EXPENSIFY_CARD,
+    DEEP_DIVE_PER_DIEM,
+    DELEGATE_ROLE_HELP_DOT_ARTICLE_LINK,
+    DENIED_CAMERA_ACCESS_INSTRUCTIONS_URL,
+    DOMAIN_VERIFICATION_HELP_URL,
+    ELECTRONIC_DISCLOSURES_URL,
+    EMAIL,
+    ENABLE_GLOBAL_REIMBURSEMENT_HELP_URL,
+    EXAMPLE_PHONE_NUMBER,
+    FEES_URL,
+    FORMATTED_EXAMPLE_PHONE_NUMBER,
+    HELP_DOC_LINKS,
+    IMPORT_TAGS_EXPENSIFY_URL,
+    IMPORT_TAGS_EXPENSIFY_URL_DEPENDENT_TAGS,
+    NEW_EXPENSIFY_URL,
+    OLD_DOT_PUBLIC_URLS,
+    ONFIDO_FACIAL_SCAN_POLICY_URL,
+    ONFIDO_PRIVACY_POLICY_URL,
+    ONFIDO_TERMS_OF_SERVICE_URL,
+    PERSONAL_DATA_PROTECTION_INFO_URL,
+    PRICING,
+    SAML_HELP_URL,
+    SELECT_WORKFLOWS_HELP_URL,
+    SET_NOTIFICATION_LINK,
+    TERMS,
+    TRAVEL_TERMS_URL,
+    UBER_TERMS_LINK,
+    UNLOCK_BANK_ACCOUNT_HELP_URL,
+    ALL_COUNTRIES,
+    CERTINIA_EXPORT_DATE,
+    CERTINIA_EXPORT_STATUS,
+    CERTINIA_MAPPING_VALUE,
+    CERTINIA_PARENT_TAG_MAPPING,
+    CERTINIA_REPORT_EXPORT_STATUS,
+    COMPANY_CARDS,
+    COUNTRY,
+    CUSTOM_UNITS,
+    DELEGATE,
+    DELEGATE_ROLE,
+    DOT_SEPARATOR,
+    DUALENTRY_EXPORT_DATE,
+    DUALENTRY_EXPORT_NON_REIMBURSABLE,
+    DUALENTRY_EXPORT_REIMBURSABLE,
+    EXIT_SURVEY,
+    IMPORT_SPREADSHEET,
+    INTEGRATION_ENTITY_MAP_TYPES,
+    INTRO_CHOICES,
+    IOU,
+    NETSUITE_EXPORT_DATE,
+    NETSUITE_EXPORT_DESTINATION,
+    NETSUITE_IMPORT,
+    NETSUITE_INVOICE_ITEM_PREFERENCE,
+    NETSUITE_JOURNALS_APPROVAL_LEVEL,
+    NETSUITE_JOURNAL_POSTING_PREFERENCE,
+    NETSUITE_REPORTS_APPROVAL_LEVEL,
+    NETSUITE_VENDOR_BILLS_APPROVAL_LEVEL,
+    NEXT_STEP,
+    MODIFIED_AMOUNT_VIOLATION_DATA,
+    ONBOARDING_ACCOUNTING_MAPPING,
+    ONBOARDING_CHOICES,
+    ONBOARDING_COMPANY_SIZE,
+    ONBOARDING_PERSONAL_TRACK_GOALS,
+    ONBOARDING_TASK_TYPE,
+    POLICY,
+    QUICKBOOKS_DESKTOP_NON_REIMBURSABLE_EXPORT_ACCOUNT_TYPE,
+    QUICKBOOKS_DESKTOP_REIMBURSABLE_ACCOUNT_TYPE,
+    QUICKBOOKS_EXPORT_DATE,
+    QUICKBOOKS_NON_REIMBURSABLE_EXPORT_ACCOUNT_TYPE,
+    QUICKBOOKS_REIMBURSABLE_ACCOUNT_TYPE,
+    REFERRAL_PROGRAM,
+    REPORT,
+    REPORT_VIOLATIONS,
+    RILLET_EXPORT_DATE,
+    RILLET_EXPORT_NON_REIMBURSABLE,
+    RILLET_EXPORT_REIMBURSABLE,
+    RTER_VIOLATION_TYPES,
+    SAGE_INTACCT_CONFIG,
+    SAGE_INTACCT_EXPORT_DATE,
+    SAGE_INTACCT_NON_REIMBURSABLE_EXPENSE_TYPE,
+    SAGE_INTACCT_REIMBURSABLE_EXPENSE_TYPE,
+    SEARCH,
+    SPEND_RULES,
+    UPGRADE_FEATURE_INTRO_MAPPING,
+    WALLET,
+    XERO_CONFIG,
+    XERO_EXPORT_DATE,
+} as const;
+
+export default CONST;
