+++ src/CONST/language/integrations/Certinia.ts
@@ -1,0 +1,31 @@
+const CERTINIA_EXPORT_STATUS = {
+    COMPLETE: 'Complete',
+    IN_PROGRESS: 'In Progress',
+    APPROVED: 'Approved',
+    SUBMITTED: 'Submitted',
+} as const;
+
+const CERTINIA_REPORT_EXPORT_STATUS = {
+    APPROVED: 'Approved',
+    SUBMITTED: 'Submitted',
+} as const;
+
+const CERTINIA_EXPORT_DATE = {
+    LAST_EXPENSE: 'LAST_EXPENSE',
+    REPORT_SUBMITTED: 'REPORT_SUBMITTED',
+    REPORT_EXPORTED: 'REPORT_EXPORTED',
+} as const;
+
+const CERTINIA_MAPPING_VALUE = {
+    DEFAULT: 'DEFAULT',
+    TAG: 'TAG',
+    REPORT_FIELD: 'REPORT_FIELD',
+} as const;
+
+const CERTINIA_PARENT_TAG_MAPPING = {
+    PARENT_TAG_PROJECTS_AND_ASSIGNMENTS: 'PROJECTS_AND_ASSIGNMENTS',
+    PARENT_TAG_PROJECTS: 'PROJECTS',
+    PARENT_TAG_ASSIGNMENTS: 'ASSIGNMENTS',
+} as const;
+
+export {CERTINIA_EXPORT_DATE, CERTINIA_EXPORT_STATUS, CERTINIA_MAPPING_VALUE, CERTINIA_PARENT_TAG_MAPPING, CERTINIA_REPORT_EXPORT_STATUS};
