+++ src/CONST/language/integrations/DualEntry.ts
@@ -1,0 +1,15 @@
+const DUALENTRY_EXPORT_REIMBURSABLE = {
+    VENDOR_BILL: 'VENDOR_BILL',
+} as const;
+
+const DUALENTRY_EXPORT_NON_REIMBURSABLE = {
+    DIRECT_EXPENSE: 'DIRECT_EXPENSE',
+} as const;
+
+const DUALENTRY_EXPORT_DATE = {
+    LAST_EXPENSE: 'LAST_EXPENSE',
+    REPORT_EXPORTED: 'REPORT_EXPORTED',
+    REPORT_SUBMITTED: 'REPORT_SUBMITTED',
+} as const;
+
+export {DUALENTRY_EXPORT_DATE, DUALENTRY_EXPORT_NON_REIMBURSABLE, DUALENTRY_EXPORT_REIMBURSABLE};
