+++ src/CONST/language/integrations/NetSuite.ts
@@ -1,0 +1,58 @@
+const NETSUITE_IMPORT = {
+    HELP_LINKS: {
+        CUSTOM_SEGMENTS: 'https://help.expensify.com/articles/expensify-classic/integrations/accounting-integrations/NetSuite#custom-segments',
+        CUSTOM_LISTS: 'https://help.expensify.com/articles/expensify-classic/integrations/accounting-integrations/NetSuite#custom-lists',
+    },
+} as const;
+
+const NETSUITE_EXPORT_DATE = {
+    LAST_EXPENSE: 'LAST_EXPENSE',
+    EXPORTED: 'EXPORTED',
+    SUBMITTED: 'SUBMITTED',
+} as const;
+
+const NETSUITE_EXPORT_DESTINATION = {
+    EXPENSE_REPORT: 'EXPENSE_REPORT',
+    VENDOR_BILL: 'VENDOR_BILL',
+    JOURNAL_ENTRY: 'JOURNAL_ENTRY',
+} as const;
+
+const NETSUITE_INVOICE_ITEM_PREFERENCE = {
+    CREATE: 'create',
+    SELECT: 'select',
+} as const;
+
+const NETSUITE_JOURNAL_POSTING_PREFERENCE = {
+    JOURNALS_POSTING_INDIVIDUAL_LINE: 'JOURNALS_POSTING_INDIVIDUAL_LINE',
+    JOURNALS_POSTING_TOTAL_LINE: 'JOURNALS_POSTING_TOTAL_LINE',
+} as const;
+
+const NETSUITE_REPORTS_APPROVAL_LEVEL = {
+    REPORTS_APPROVED_NONE: 'REPORTS_APPROVED_NONE',
+    REPORTS_SUPERVISOR_APPROVED: 'REPORTS_SUPERVISOR_APPROVED',
+    REPORTS_ACCOUNTING_APPROVED: 'REPORTS_ACCOUNTING_APPROVED',
+    REPORTS_APPROVED_BOTH: 'REPORTS_APPROVED_BOTH',
+} as const;
+
+const NETSUITE_VENDOR_BILLS_APPROVAL_LEVEL = {
+    VENDOR_BILLS_APPROVED_NONE: 'VENDOR_BILLS_APPROVED_NONE',
+    VENDOR_BILLS_APPROVAL_PENDING: 'VENDOR_BILLS_APPROVAL_PENDING',
+    VENDOR_BILLS_APPROVED: 'VENDOR_BILLS_APPROVED',
+} as const;
+
+const NETSUITE_JOURNALS_APPROVAL_LEVEL = {
+    JOURNALS_APPROVED_NONE: 'JOURNALS_APPROVED_NONE',
+    JOURNALS_APPROVAL_PENDING: 'JOURNALS_APPROVAL_PENDING',
+    JOURNALS_APPROVED: 'JOURNALS_APPROVED',
+} as const;
+
+export {
+    NETSUITE_EXPORT_DATE,
+    NETSUITE_EXPORT_DESTINATION,
+    NETSUITE_IMPORT,
+    NETSUITE_INVOICE_ITEM_PREFERENCE,
+    NETSUITE_JOURNALS_APPROVAL_LEVEL,
+    NETSUITE_JOURNAL_POSTING_PREFERENCE,
+    NETSUITE_REPORTS_APPROVAL_LEVEL,
+    NETSUITE_VENDOR_BILLS_APPROVAL_LEVEL,
+};
