+++ src/CONST/language/integrations/QuickBooks.ts
@@ -1,0 +1,37 @@
+const QUICKBOOKS_REIMBURSABLE_ACCOUNT_TYPE = {
+    VENDOR_BILL: 'bill',
+    CHECK: 'check',
+    JOURNAL_ENTRY: 'journal_entry',
+} as const;
+
+const QUICKBOOKS_DESKTOP_REIMBURSABLE_ACCOUNT_TYPE = {
+    VENDOR_BILL: 'VENDOR_BILL',
+    CHECK: 'CHECK',
+    JOURNAL_ENTRY: 'JOURNAL_ENTRY',
+} as const;
+
+const QUICKBOOKS_EXPORT_DATE = {
+    LAST_EXPENSE: 'LAST_EXPENSE',
+    REPORT_EXPORTED: 'REPORT_EXPORTED',
+    REPORT_SUBMITTED: 'REPORT_SUBMITTED',
+} as const;
+
+const QUICKBOOKS_NON_REIMBURSABLE_EXPORT_ACCOUNT_TYPE = {
+    CREDIT_CARD: 'credit_card',
+    DEBIT_CARD: 'debit_card',
+    VENDOR_BILL: 'bill',
+} as const;
+
+const QUICKBOOKS_DESKTOP_NON_REIMBURSABLE_EXPORT_ACCOUNT_TYPE = {
+    CREDIT_CARD: 'CREDIT_CARD_CHARGE',
+    CHECK: 'CHECK',
+    VENDOR_BILL: 'VENDOR_BILL',
+} as const;
+
+export {
+    QUICKBOOKS_DESKTOP_NON_REIMBURSABLE_EXPORT_ACCOUNT_TYPE,
+    QUICKBOOKS_DESKTOP_REIMBURSABLE_ACCOUNT_TYPE,
+    QUICKBOOKS_EXPORT_DATE,
+    QUICKBOOKS_NON_REIMBURSABLE_EXPORT_ACCOUNT_TYPE,
+    QUICKBOOKS_REIMBURSABLE_ACCOUNT_TYPE,
+};
