+++ src/CONST/language/integrations/Rillet.ts
@@ -1,0 +1,15 @@
+const RILLET_EXPORT_REIMBURSABLE = {
+    VENDOR_BILL: 'VENDOR_BILL',
+} as const;
+
+const RILLET_EXPORT_NON_REIMBURSABLE = {
+    CREDIT_CARD_CHARGE: 'CREDIT_CARD_CHARGE',
+} as const;
+
+const RILLET_EXPORT_DATE = {
+    LAST_EXPENSE: 'LAST_EXPENSE',
+    REPORT_EXPORTED: 'REPORT_EXPORTED',
+    REPORT_SUBMITTED: 'REPORT_SUBMITTED',
+} as const;
+
+export {RILLET_EXPORT_DATE, RILLET_EXPORT_NON_REIMBURSABLE, RILLET_EXPORT_REIMBURSABLE};
