+++ src/CONST/language/integrations/SageIntacct.ts
@@ -1,0 +1,27 @@
+const SAGE_INTACCT_CONFIG = {
+    MAPPINGS: {
+        DEPARTMENTS: 'departments',
+        CLASSES: 'classes',
+        LOCATIONS: 'locations',
+        CUSTOMERS: 'customers',
+        PROJECTS: 'projects',
+    },
+} as const;
+
+const SAGE_INTACCT_REIMBURSABLE_EXPENSE_TYPE = {
+    EXPENSE_REPORT: 'EXPENSE_REPORT',
+    VENDOR_BILL: 'VENDOR_BILL',
+} as const;
+
+const SAGE_INTACCT_NON_REIMBURSABLE_EXPENSE_TYPE = {
+    CREDIT_CARD_CHARGE: 'CREDIT_CARD_CHARGE',
+    VENDOR_BILL: 'VENDOR_BILL',
+} as const;
+
+const SAGE_INTACCT_EXPORT_DATE = {
+    LAST_EXPENSE: 'LAST_EXPENSE',
+    EXPORTED: 'EXPORTED',
+    SUBMITTED: 'SUBMITTED',
+} as const;
+
+export {SAGE_INTACCT_CONFIG, SAGE_INTACCT_EXPORT_DATE, SAGE_INTACCT_NON_REIMBURSABLE_EXPENSE_TYPE, SAGE_INTACCT_REIMBURSABLE_EXPENSE_TYPE};
