+++ src/CONST/language/integrations/Xero.ts
@@ -1,0 +1,20 @@
+const XERO_CONFIG = {
+    INVOICE_STATUS: {
+        DRAFT: 'DRAFT',
+        AWAITING_APPROVAL: 'AWT_APPROVAL',
+        AWAITING_PAYMENT: 'AWT_PAYMENT',
+    },
+    TRACKING_CATEGORY_OPTIONS: {
+        DEFAULT: 'DEFAULT',
+        TAG: 'TAG',
+        REPORT_FIELD: 'REPORT_FIELD',
+    },
+} as const;
+
+const XERO_EXPORT_DATE = {
+    LAST_EXPENSE: 'LAST_EXPENSE',
+    REPORT_EXPORTED: 'REPORT_EXPORTED',
+    REPORT_SUBMITTED: 'REPORT_SUBMITTED',
+} as const;
+
+export {XERO_CONFIG, XERO_EXPORT_DATE};
