+++ src/languages/types/Country.ts
@@ -1,0 +1,2 @@
+// eslint-disable-next-line import/prefer-default-export -- Matches the named type import used by language files.
+export type {Country} from '@src/CONST/language/Countries';
