+++ src/languages/types/OriginalMessage.ts
@@ -1,0 +1,1 @@
+export type {OriginalMessageReportPreview, OriginalMessageSettlementAccountLocked, PersonalRulesModifiedFields, PolicyRulesModifiedFields} from '@src/types/onyx/OriginalMessageCore';
