+++ src/languages/types/Policy.ts
@@ -1,0 +1,1 @@
+export type {AllConnectionName, ConnectionName, PolicyConnectionSyncStage, SageIntacctMappingName} from '@src/types/onyx/PolicyCore';
