+++ src/languages/types/TransactionViolation.ts
@@ -1,0 +1,1 @@
+export type {DelegateRole, ViolationDataType} from '@src/types/onyx/TransactionViolationCore';
