+++ src/languages/types/params.ts
@@ -1,0 +1,14 @@
+import type {ConnectionName} from './Policy';
+
+type ExportAgainModalDescriptionParams = {
+    reportName: string;
+    connectionName: ConnectionName;
+    connectionNameFriendly?: string;
+};
+
+type ExportIntegrationSelectedParams = {
+    connectionName: ConnectionName;
+    connectionNameFriendly?: string;
+};
+
+export type {ExportAgainModalDescriptionParams, ExportIntegrationSelectedParams};
