+++ src/libs/StringUtils/startsWithVowel.ts
@@ -1,0 +1,6 @@
+/** Returns whether a string begins with an English vowel. */
+function startsWithVowel(str: string): boolean {
+    return /^[aeiouAEIOU]/.test(str);
+}
+
+export default startsWithVowel;
