+++ src/types/onyx/Account.ts
@@ -4,9 +4,7 @@ import type {ValueOf} from 'type-fest';
 
 import type DismissedReferralBanners from './DismissedReferralBanners';
 import type * as OnyxCommon from './OnyxCommon';
-
-/** The role of the delegate */
-type DelegateRole = ValueOf<typeof CONST.DELEGATE_ROLE>;
+import type {DelegateRole} from './TransactionViolationCore';
 
 /** Model of delegate */
 type Delegate = OnyxCommon.OnyxValueWithOfflineFeedback<{
