+++ src/types/onyx/ExpenseRule.ts
@@ -1,3 +1,4 @@
+import type {PolicyRuleTaxRate} from './ExpenseRuleCore';
 import type * as OnyxCommon from './OnyxCommon';
 
 /** Model of tax rate on a personal rule */
@@ -7,18 +8,6 @@ type PersonalRuleTaxRate = {
 
     /** The tax rate */
     value: string;
-};
-
-/** Model of tax rate on a policy rule */
-type PolicyRuleTaxRate = {
-    /** The tax ID */
-    externalID: string;
-
-    /** The tax rate */
-    value: string;
-
-    /** The name of the tax rate */
-    name: string;
 };
 
 /** Model of expense rule */
