+++ src/types/onyx/OriginalMessage.ts
@@ -4,9 +4,9 @@ import type DeepValueOf from '@src/types/utils/DeepValueOf';
 import type {ValueOf} from 'type-fest';
 
 import type {CardID} from './Card';
-import type {PolicyRuleTaxRate} from './ExpenseRule';
 import type {Attendee} from './IOU';
 import type {OldDotOriginalMessageMap} from './OldDotAction';
+import type {OriginalMessageReportPreview, OriginalMessageSettlementAccountLocked, PersonalRulesModifiedFields, PolicyRulesModifiedFields} from './OriginalMessageCore';
 import type {AllConnectionName} from './Policy';
 import type {PolicyChangeLogCopyReportActionNames} from './ReportAction';
 import type ReportActionName from './ReportActionName';
@@ -371,15 +371,6 @@ type ChronosOOOEvent = {
 type OriginalMessageChronosOOOList = {
     /** Collection of OOO events to show in report action */
     events: ChronosOOOEvent[];
-};
-
-/** Model of `report preview` report action */
-type OriginalMessageReportPreview = {
-    /** ID of the report to be previewed */
-    linkedReportID: string;
-
-    /** Collection of accountIDs of users mentioned in report */
-    whisperedTo?: number[];
 };
 
 /** Possible values of policy budget frequency */
@@ -1077,71 +1068,6 @@ type OriginalMessageConciergeAutoMatchVendor = {
     reasoning?: string;
 };
 
-/** Policy rules modified fields */
-type PolicyRulesModifiedFields = {
-    /** The value that the merchant was changed to */
-    merchant?: string;
-
-    /** The value that the amount was changed to */
-    category?: string;
-
-    /** The value that the tag was changed to */
-    tag?: string;
-
-    /** The value that the description was changed to (backend uses "comment" key) */
-    comment?: string;
-
-    /** The value that the description was changed to (display key, mapped from "comment") */
-    description?: string;
-
-    /** The value that the billable status was changed to */
-    billable?: boolean;
-
-    /** The value that the reimbursable status was changed to */
-    reimbursable?: boolean;
-
-    /** The value that the tax was changed to */
-    tax?: {
-        /** The tax rate being used  */
-        // eslint-disable-next-line @typescript-eslint/naming-convention
-        field_id_TAX: PolicyRuleTaxRate;
-    };
-};
-
-/** Personal rules modified fields */
-type PersonalRulesModifiedFields = {
-    /** The value that the merchant was changed to */
-    merchant?: string;
-
-    /** The value that the amount was changed to */
-    category?: string;
-
-    /** The value that the tag was changed to */
-    tag?: string;
-
-    /** The value that the description was changed to (backend uses "comment" key) */
-    comment?: string;
-
-    /** The value that the description was changed to (display key, mapped from "comment") */
-    description?: string;
-
-    /** The value that the billable status was changed to */
-    billable?: boolean;
-
-    /** The value that the reimbursable status was changed to */
-    reimbursable?: boolean;
-
-    /** The value that the tax was changed to */
-    tax?: {
-        /** The tax rate being used  */
-        // eslint-disable-next-line @typescript-eslint/naming-convention
-        field_id_TAX: PolicyRuleTaxRate;
-    };
-
-    /** The value that the report name was set to */
-    reportName?: string;
-};
-
 /** Model of a `travel update` report action */
 type OriginalMessageTravelUpdate = Reservation & UpdateOperationType;
 
@@ -1671,17 +1597,6 @@ type OriginalMessageDelegateSubmit = {
 type OriginalMessageActionableCard3DSTransactionApproval = TransactionPending3DSReview;
 
 /**
- * Model of settlement account locked report action
- */
-type OriginalMessageSettlementAccountLocked = {
-    /** The masked bank account number that was locked */
-    maskedBankAccountNumber: string;
-
-    /** The policy the bank account is connected to and is being notified */
-    policyID: string;
-};
-
-/**
  * Original message for Expensify Card issue/replacement actions
  */
 type IssueNewCardOriginalMessage = OriginalMessage<
@@ -1854,6 +1769,7 @@ export type {
     ChronosOOOEvent,
     PaymentMethodType,
     OriginalMessageSource,
+    OriginalMessageReportPreview,
     Decision,
     PolicyRulesModifiedFields,
     PersonalRulesModifiedFields,
