+++ src/types/onyx/ReportActionName.ts
@@ -1,7 +1,3 @@
-import type CONST from '@src/CONST';
-import type DeepValueOf from '@src/types/utils/DeepValueOf';
-
-/** The name (or type) of a reportAction */
-type ReportActionName = DeepValueOf<typeof CONST.REPORT.ACTIONS.TYPE>;
+import type {ReportActionName} from './ReportActionCore';
 
 export default ReportActionName;
